requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
pandas==2.2.2
duckdb==1.0.0
numpy==1.26.4
python-dotenv==1.0.1
boto3==1.34.162
requests==2.32.3
pydantic==2.8.2
